requirements.txt
AdvancedHTTPServer==2.2.0
aiodns==2.0.0
aiohttp==3.6.2
aiomultiprocess==0.8.0
aiosqlite==0.15.0
ajpy==0.0.4
alembic==1.4.3.dev0
aniso8601==8.0.0
apispec==3.3.1
apispec-webframeworks==0.5.2
appdirs==1.4.4
asgiref==3.2.10
asn1crypto==1.4.0
async-timeout==3.0.1
atomicwrites==1.4.0
attrs==20.3.0
autobahn==17.10.1
Automat==20.2.0
Babel==2.8.0
backcall==0.2.0
backdoor-factory==0.0.0
basemap==1.2.1
bcrypt==3.1.7
beautifulsoup4==4.9.3
binwalk==2.2.0
blinker==1.4
boltons==19.1.0
bottle==0.12.15
Brlapi==0.8.0
Brotli==1.0.9
cairocffi==0.9.0
capstone==3.0.5
cbor==1.0.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
chirp==0.3.0.dev0
chrome-gnome-shell==0.0.0
click==7.1.2
click-plugins==1.1.1
colorama==0.4.3
configobj==5.0.6
constantly==15.1.0
crcelk==1.3
cryptography==3.2.1
cupshelpers==1.0
cycler==0.10.0
Cython==0.29.21
dbus-python==1.2.16
debtags==2.1
decorator==4.4.2
deprecation==2.0.7
dicttoxml==1.7.4
distro==1.5.0
distro-info==0.24
Django==3.1.2
django-cors-headers==3.5.0
django-rest-auth==0.9.5
django-taggit==1.3.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
dnslib==0.9.14
dnspython==2.0.0
docutils==0.16
ecdsa==0.16.0
email-validator==1.1.1
entrypoints==0.3
faraday-client==1.0.0
faraday-plugins==1.2
faradaysec==3.12
feedparser==5.2.1
fierce==1.4.0
filedepot==0.5.2
filteralchemy==0.1.0
flasgger==0.9.5
Flask==1.1.2
Flask-BabelEx==0.9.4
Flask-Classful==0.14.1
Flask-KVSession-fork==0.6.3
Flask-Login==0.5.0
Flask-Mail==0.9.1
Flask-Principal==0.4.0
Flask-RESTful==0.3.8
Flask-Security-Too==3.4.2
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.3
freegames==2.3.2
future==0.18.2
GeoIP==1.3.2
geoip2==2.9.0
geojson==2.5.0
graphene==2.1.7
graphene-sqlalchemy==2.1.2
graphql-core==2.2.1
graphql-relay==2.0.0
h11==0.11.0
h2==3.2.0
hashID==3.1.4
hpack==3.0.0
html2text==2020.1.16
html5lib==1.1
httplib2==0.18.1
humanize==0.0.0
hupper==1.10.2
hyperframe==5.2.0
hyperlink==19.0.0
icalendar==4.0.3
idna==2.10
impacket==0.9.21
importlib-metadata==1.6.0
incremental==17.5.0
invoke==1.4.1
IPy==1.0
ipython==7.19.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.17.0
jeepney==0.5.0
Jinja2==2.11.2
jsonschema==3.2.0
jupyter-core==4.6.3
kaitaistruct==0.8
keyring==21.5.0
keyrings.alt==4.0.1
KismetCaptureFreaklabsZigbee==2018.7.0
KismetCaptureRtl433==2019.9.1
KismetCaptureRtladsb==2019.10.1
KismetCaptureRtlamr==2019.10.1
kiwisolver==1.3.1
ldap3==2.7
ldapdomaindump==0.9.3
louis==3.15.0
lxml==4.6.1
lz4==3.1.0+dfsg
Mako==1.1.3
Markdown==3.3.3
MarkupSafe==1.1.1
marshmallow==3.7.1
marshmallow-sqlalchemy==0.19.0
matplotlib==3.3.2
maxminddb==1.4.1
mechanize==0.4.5
mercurial==5.5.2
mistune==0.8.4
mitmproxy==5.1.1
more-itertools==4.2.0
MouseInfo==0.1.3
msgpack==0.6.2
multidict==5.0.0
mysqlclient==1.4.4
nassl==3.1.0
nbformat==5.0.8
netaddr==0.7.19
nplusone==1.0.0
numpy==1.19.4
olefile==0.46
packaging==20.4
paramiko==2.7.2
parso==0.7.0
passlib==1.7.2
Paste==3.5.0
PasteDeploy==2.1.1
PasteScript==2.0.2
patator==0.9
pcapy==0.11.4
pefile==2019.4.18
pexpect==4.6.0
pgspecial==1.11.10
pickleshare==0.7.5
Pillow==8.0.1
plaster==1.0
plaster-pastedeploy==0.5
plecost==1.1.2
plotly==4.12.0
pluggy==0.13.0
pluginbase==1.0.0
ply==3.11
prettytable==0.7.2
promise==2.2
prompt-toolkit==3.0.8
protobuf==3.12.3
psycopg2==2.8.5
publicsuffix2==2.20191221
py==1.9.0
py-ubjson==0.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
PyAutoGUI==0.9.52
pycairo==1.16.2
pycares==3.1.1
pycparser==2.20
pycrypto==2.6.1
pycryptodomex==3.9.7
pycups==2.0.1
pycurl==7.43.0.6
pydot==1.4.1
pyee==7.0.2
pyenchant==3.0.1
PyGetWindow==0.0.9
Pygments==2.3.1
PyGObject==3.38.0
PyHamcrest==1.9.0
pyinotify==0.9.6
PyJWT==1.7.1
PyMsgBox==1.0.9
pymssql==2.1.4
PyNaCl==1.4.0
PyOpenGL==3.1.5
pyOpenSSL==19.1.0
pyotp==2.3.0
pyparsing==2.4.7
pyperclip==1.8.0
pypng==0.0.20
pyppeteer==0.2.2
pyproj==3.0.0
PyQRCode==1.2.1
PyQt5==5.15.1
PyQt5-sip==12.8.1
pyramid==1.10.4
PyRect==0.1.4
pyrsistent==0.15.5
PyScreeze==0.1.26
pyserial==3.5b0
pyshp==2.1.2
PySimpleSOAP==1.16.2
pysmi==0.3.2
pysnmp==4.4.12
PySocks==1.7.1
pytest==4.6.11
python-apt==2.1.5
python-dateutil==2.8.1
python-debian==0.1.38
python-debianbts==3.0.2
python-editor==1.0.3
python-magic==0.4.16
python-magic-ahupp==0.4.13
python-pam==1.8.4
python-snappy==0.5.3
python3-xlib==0.15
PyTrie==0.2
PyTweening==1.0.3
pytz==2020.4
pyxdg==0.26
PyYAML==5.3.1
qrcode==6.1
redis==3.3.11
reportbug==7.8.0
requests==2.24.0
requests-file==1.5.1
requests-toolbelt==0.8.0
retrying==1.3.3
roman==2.0.0
rq==1.5.2
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
rule-engine==1.1.0
Rx==1.6.1
scipy==1.5.3
SecretStorage==3.2.0
selenium==4.0.0a1
service-identity==18.1.0
setproctitle==1.1.10
shodan==1.24.0
simplejson==3.17.0
simplekv==0.13.0
sip==4.19.24
six==1.15.0
smoke-zephyr==2.0.0
sortedcontainers==2.1.0
soupsieve==2.0.1
speaklater==1.4
SQLAlchemy==1.3.20
sqlalchemy-schemadisplay==1.3
sqlparse==0.3.1
sslyze==3.1.0
syslog-rfc5424-formatter==1.2.2
tabulate==0.8.7
Tempita==0.5.2
termcolor==1.1.0
terminaltables==3.1.0
termineter==1.0.4
texttable==1.6.3
theHarvester==3.2.0
tld==0.11.11
tls-parser==1.2.2
tornado==6.0.4
tqdm==4.51.0
traitlets==5.0.5
translationstring==1.4
Twisted==20.3.0
txaio==20.4.1
typing-extensions==3.7.4.3
tzlocal==2.1
u-msgpack-python==2.3.0
unattended-upgrades==0.1
unicodecsv==0.14.1
Unidecode==1.1.1
urllib3==1.25.9
urwid==2.1.1
uvloop==0.14.0
venusian==3.0.0
vinetto==0.8.0
wafw00f==2.1.0
wapiti3==3.0.3
wcwidth==0.1.9
webargs==6.1.0
webencodings==0.5.1
WebOb==1.8.6
websocket-client==0.57.0
websockets==8.1
Werkzeug==1.0.1
wfuzz==3.0.1
wifite==2.5.7
wsaccel==0.6.2
wsproto==0.15.0
WTForms==2.2.1
wxPython==4.0.7
xcffib==0.8.1
XlsxWriter==1.1.2
yara-python==4.0.2
yarl==1.4.2
yaswfp==0.9.3
zim==0.73.3
zipp==1.0.0
zope.deprecation==4.4.0
zope.interface==5.1.2
